Rename truffle-artifactor to @truffle/artifactor
