+++ packages/artifactor/index.js
