+++ packages/artifactor/test/customoptions.js
