+++ packages/debugger/test/helpers.js
@@ -5,7 +5,7 @@ import path from "path";
 import fs from "fs-extra";
 import Contracts from "@truffle/workflow-compile";
 import Debug from "@truffle/debug-utils";
-import Artifactor from "truffle-artifactor";
+import Artifactor from "@truffle/artifactor";
 import Web3 from "web3";
 import Migrate from "truffle-migrate";
 import Box from "@truffle/box";
