+++ packages/truffle-core/lib/commands/migrate.js
@@ -142,7 +142,7 @@ const command = {
   },
 
   prepareConfigForRealMigrations: async function(buildDir, options) {
-    const Artifactor = require("truffle-artifactor");
+    const Artifactor = require("@truffle/artifactor");
     const Resolver = require("truffle-resolver");
     const Migrate = require("truffle-migrate");
     const Config = require("truffle-config");
@@ -175,7 +175,7 @@ const command = {
   },
 
   run: function(options, done) {
-    const Artifactor = require("truffle-artifactor");
+    const Artifactor = require("@truffle/artifactor");
     const Resolver = require("truffle-resolver");
     const Migrate = require("truffle-migrate");
     const Contracts = require("@truffle/workflow-compile");
