+++ packages/truffle-core/lib/commands/test.js
@@ -46,7 +46,7 @@ const command = {
     const dir = require("node-dir");
     const temp = require("temp").track();
     const Config = require("truffle-config");
-    const Artifactor = require("truffle-artifactor");
+    const Artifactor = require("@truffle/artifactor");
     const Test = require("../test");
     const fs = require("fs");
     const { promisify } = require("util");
