+++ packages/truffle-core/package.json
@@ -49,7 +49,7 @@
     "spawn-args": "^0.1.0",
     "temp": "^0.8.3",
     "@truffle/compile-solidity": "^4.2.3",
-    "truffle-artifactor": "^4.0.30",
+    "@truffle/artifactor": "^4.0.30",
     "truffle-config": "^1.1.20",
     "truffle-contract": "^4.0.31",
     "truffle-contract-sources": "^0.1.5",
