+++ packages/truffle-core/test/config.js
@@ -6,7 +6,7 @@ var Contracts = require("@truffle/workflow-compile");
 var Ganache = require("ganache-core");
 var provision = require("@truffle/provisioner");
 var Resolver = require("truffle-resolver");
-var Artifactor = require("truffle-artifactor");
+var Artifactor = require("@truffle/artifactor");
 
 describe("config", function() {
   var config;
