+++ packages/truffle-core/test/lib/commands/compile.js
@@ -1,7 +1,7 @@
 var assert = require("chai").assert;
 var Box = require("@truffle/box");
 var Contracts = require("@truffle/workflow-compile");
-var Artifactor = require("truffle-artifactor");
+var Artifactor = require("@truffle/artifactor");
 var Resolver = require("truffle-resolver");
 var MemoryStream = require("memorystream");
 var command = require("../../../lib/commands/compile");
