+++ packages/truffle-core/test/migrate.js
@@ -8,7 +8,7 @@ var fs = require("fs-extra");
 var glob = require("glob");
 var Ganache = require("ganache-core");
 var Resolver = require("truffle-resolver");
-var Artifactor = require("truffle-artifactor");
+var Artifactor = require("@truffle/artifactor");
 var Web3 = require("web3");
 
 describe("migrate", function() {
