+++ packages/truffle-core/test/npm.js
@@ -6,7 +6,7 @@ var path = require("path");
 var mkdirp = require("mkdirp");
 var async = require("async");
 var Resolver = require("truffle-resolver");
-var Artifactor = require("truffle-artifactor");
+var Artifactor = require("@truffle/artifactor");
 var Contracts = require("@truffle/workflow-compile");
 
 describe("NPM integration", function() {
