+++ packages/truffle-environment/package.json
@@ -16,7 +16,7 @@
     "@truffle/expect": "^0.0.10",
     "ganache-core": "2.7.0",
     "node-ipc": "^9.1.1",
-    "truffle-artifactor": "^4.0.30",
+    "@truffle/artifactor": "^4.0.30",
     "truffle-interface-adapter": "^0.2.5",
     "truffle-resolver": "^5.0.16",
     "web3": "1.2.1"
