+++ packages/workflow-compile/package.json
@@ -9,7 +9,7 @@
     "@truffle/compile-solidity": "^4.2.3",
     "@truffle/compile-vyper": "^1.0.27",
     "@truffle/external-compile": "^1.0.16",
-    "truffle-artifactor": "^4.0.30",
+    "@truffle/artifactor": "^4.0.30",
     "truffle-config": "^1.1.20",
     "truffle-resolver": "^5.0.16"
   },
